epsg double => int に変更

+++ src/c_wrapper/geo_reference_c.cpp
@@ -49,7 +49,7 @@ extern "C" {
                 GeoReference,
                 TVec3d,
                 handle->convert(point, convert_axis, epsg),
-                , TVec3d point, bool convert_axis, double epsg)
+                , TVec3d point, bool convert_axis, int epsg)
 
     DLL_VALUE_FUNC(plateau_geo_reference_get_reference_point,
                   GeoReference,
